Backend/requirements.txt
# Core ML/DL packages
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data science and analysis
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Image processing
Pillow>=9.0.0

# Utilities
tqdm>=4.64.0

# Grad-CAM visualization
torchcam>=0.3.0

# Optional: For Jupyter notebooks (if needed)
# jupyter>=1.0.0
# ipykernel>=6.0.0